requirements.txt
fastapi
uvicorn
aiogram
google-generativeai
supabase
python-dotenv
httpx
